+++ src/helpers/effects/chat-message-effect.ts
@@ -1,6 +1,8 @@
-import { Effect, Firebot } from 'firebot-custom-scripts-types';
+import { Effect, Firebot, RunRequest } from 'firebot-custom-scripts-types';
+import { CustomEffect } from './custom-effect';
+import { ScriptParams } from '../../main';
 
-export class ChatMessageEffect implements Effect {
+export class ChatMessageEffect implements Effect, CustomEffect {
     readonly type: Firebot.KnownEffectType = 'firebot:chat';
 
     readonly message: string;
@@ -11,6 +13,10 @@ export class ChatMessageEffect implements Effect {
 
     toString(): string {
         return `ChatMessageEffect { message: ${this.message} }`;
+    }
+
+    async execute(runRequest: RunRequest<ScriptParams>): Promise<void> {
+        runRequest.modules.twitchChat.sendChatMessage(this.message, undefined, 'bot');
     }
 
     // eslint-disable-next-line no-undef